requirements.txt
gigachat>=0.2.0
openai>=1.0.0
requests>=2.28.0
jinja2>=3.1.0
reportlab>=4.0.0
weasyprint>=62.0
xhtml2pdf>=0.2.16
flask>=3.0.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
